refactor: Move `mergeZapperPropsWithAddressables` to its own file
